feat: complete pytorch v1

+++ trainer/pytorch/dataset.py
@@ -1,5 +1,6 @@
 from typing import Union
 
+import numpy as np
 from jina import DocumentArray
 from jina.types.arrays.memmap import DocumentArrayMemmap
 from torch.utils.data import Dataset
@@ -38,7 +39,10 @@ class JinaSiameseDataset(Dataset):
         for doc in inputs:
             for match in doc.matches:
                 self._pairs.append(
-                    ((doc.blob, match.blob), match.tags['trainer']['label'])
+                    (
+                        (doc.blob.astype(np.float32), match.blob.astype(np.float32)),
+                        np.float32(match.tags['trainer']['label']),
+                    )
                 )
 
     def __len__(self):
